fix: current employee not returning projects
